from repo

+++ metrics/Models/DataSourceResponseModel.cs
@@ -3,9 +3,9 @@ namespace metrics.Models
     public class DataSourceResponseModel
     {
         public object Data { get; set; }
-        public int Total { get; set; }
+        public long Total { get; set; }
 
-        public DataSourceResponseModel(object data, int total) 
+        public DataSourceResponseModel(object data, long total) 
         {
             Data = data;
             Total = total;
